+++ src/Command/UpdateCommand.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Command;
 
+use App\Service\ScreenLayoutService;
 use App\Service\TemplateService;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -21,6 +22,7 @@ class UpdateCommand extends Command
 {
     public function __construct(
         private readonly TemplateService $templateService,
+        private readonly ScreenLayoutService $screenLayoutService,
         ?string $name = null,
     ) {
         parent::__construct($name);
@@ -51,7 +53,7 @@ class UpdateCommand extends Command
             return Command::FAILURE;
         }
 
-        $allTemplates = $this->templateService->getAllTemplates();
+        $allTemplates = $this->templateService->getAll();
         $installedTemplates = array_filter($allTemplates, fn ($entry): bool => $entry->installed);
 
         // If no installed templates, we assume that this is a new installation and offer to install all templates.
@@ -59,7 +61,7 @@ class UpdateCommand extends Command
             $question = new ConfirmationQuestion('No templates are installed. Install all '.count($allTemplates).'?');
             $installAll = $io->askQuestion($question);
 
-            if ('yes' === $installAll) {
+            if ($installAll) {
                 $io->info('Installing all templates...');
                 $command = new ArrayInput([
                     'command' => 'app:templates:install',
@@ -71,6 +73,30 @@ class UpdateCommand extends Command
             $io->info('Updating existing template...');
             $command = new ArrayInput([
                 'command' => 'app:templates:update',
+            ]);
+            $application->doRun($command, $output);
+        }
+
+        $allScreenLayouts = $this->screenLayoutService->getAll();
+        $installedScreenLayouts = array_filter($allScreenLayouts, fn ($entry): bool => $entry->installed);
+
+        // If no installed screen layouts, we assume that this is a new installation and offer to install all screen layouts.
+        if ($isInteractive && 0 === count($installedScreenLayouts)) {
+            $question = new ConfirmationQuestion('No screen layouts are installed. Install all '.count($allScreenLayouts).'?');
+            $installAll = $io->askQuestion($question);
+
+            if ($installAll) {
+                $io->info('Installing all screen layouts...');
+                $command = new ArrayInput([
+                    'command' => 'app:screen-layouts:install',
+                    '--all' => true,
+                ]);
+                $application->doRun($command, $output);
+            }
+        } else {
+            $io->info('Updating existing screen layouts...');
+            $command = new ArrayInput([
+                'command' => 'app:screen-layouts:update',
             ]);
             $application->doRun($command, $output);
         }